fix: (#5) Fixed incorrect alphabetic ordering of students when generating PDF-s
This is the ugliest fix imaginable here, instead of correctly handling UTF-8 in the application, we just call back into Postgres that does it correctly, at the cost of +1 db round trip.
This makes this extremely ineffective, but for the load we present it with, this should not matter, and if it continues to run on the rpi, then it's gonna be slow whatever I do with it, so can't be arsed to properly fix it correctly.

closes #2

+++ app/controllers/groups_controller.rb
@@ -14,10 +14,14 @@ class GroupsController < ApplicationController
       format.json
       format.pdf do
         begin
+          puts ">>>: #{parse_attendance_sheet_data}"
           doc = @group.attendance_sheet(*parse_attendance_sheet_data)
+          puts ">>>: #{doc.inspect}"
           send_data doc.file.download, filename: "#{@group.name}.pdf", type: 'application/pdf'
+        rescue => e
+          logger.error("error generating group pdf: #{e.message}")
         ensure
-          doc.destroy!
+          doc.destroy! unless doc.nil?
         end
       end
     end
